% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtr.r
\name{filtr}
\alias{filtr}
\title{Filter by column
Show marginal of specific column of data. Wrapper for dplyr::select()}
\usage{
filtr(dataset, col, ...)
}
\arguments{
\item{"dataset"}{for use, column "col"}
}
\value{
one specific column
}
\description{
Filter by column
Show marginal of specific column of data. Wrapper for dplyr::select()
}
