% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstandardize.R
\name{func2}
\alias{func2}
\title{Matrix Standardization}
\usage{
func2(M)
}
\arguments{
\item{M}{matrix with multiple rows.}
}
\value{
A new, standardized matrix.
}
\description{
Standardize the columns of a matrix.
}
\examples{

func2(data3)

}
