% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/%matcalc%.R, R/grapes-matcalc-grapes.R
\name{\%matcalc\%}
\alias{\%matcalc\%}
\alias{\%matcalc\%}
\title{Calculator for  X^T A^{-1} X}
\usage{
A \%matcalc\% X

A \%matcalc\% X
}
\arguments{
\item{A}{matrix (must be invertible)}

\item{X}{numerical vector}

\item{A}{matrix (must be invertible)}

\item{X}{numerical vector}
}
\value{
calculated matrix

calculated matrix
}
\description{
Working on a matrix, A, and a numerical vector X, performs the calculation in the title line.

Working on a matrix, A, and a numerical vector X, performs the calculation in the title line.
}
\examples{

attach(data4)
a `\%matcalc\%` x

attach(data4)
a `\%matcalc\%` x
}
