% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanvarsd.R
\name{meanvarsd}
\alias{meanvarsd}
\title{Random Variable Statistic Calculator.}
\usage{
meanvarsd(d)
}
\arguments{
\item{d}{dataframe with two columns, x and p. x represents the value of the RV. p represents the probability corresponding to that value.}
}
\value{
vector of length 3 (mean, var, sd)
}
\description{
Given a discrete random variable distribution, return the mean, variance and standard deviation.
}
\examples{
meanvarsd((rnorm(10)))
}
