% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottr.r
\name{plottr}
\alias{plottr}
\title{Plot for dataset data1}
\usage{
plottr(dataset, ...)
}
\arguments{
\item{two-column}{numerical "dataset" to be plotted.}
}
\value{
plot of data
}
\description{
Constructs a scatterplot of the first two columns of a numerical data table.
}
